Move common components from `pyro` directory to `pyrocommon`
