+++ src/core/algorithms/fd/pyro/pyro.cpp
@@ -6,6 +6,8 @@
 
 #include <easylogging++.h>
 
+#include "algorithms/fd/pyrocommon/core/fd_g1_strategy.h"
+#include "algorithms/fd/pyrocommon/core/key_g1_strategy.h"
 #include "config/error/option.h"
 #include "config/max_lhs/option.h"
 #include "config/names_and_descriptions.h"
