+++ src/core/algorithms/fd/pyro/pyro.h
@@ -4,8 +4,8 @@
 #include <mutex>
 
 #include "algorithms/fd/pli_based_fd_algorithm.h"
-#include "core/dependency_consumer.h"
-#include "core/search_space.h"
+#include "algorithms/fd/pyrocommon/core/dependency_consumer.h"
+#include "algorithms/fd/pyrocommon/core/search_space.h"
 
 namespace algos {
