+++ src/core/algorithms/fd/pyrocommon/core/fd_g1_strategy.cpp
