+++ src/core/algorithms/fd/pyrocommon/core/vertical_info.h
