+++ src/core/algorithms/fd/pyrocommon/model/partial_fd.h
