+++ src/core/algorithms/fd/pyrocommon/model/pli_cache.h
@@ -6,7 +6,7 @@ class ProfilingContext;
 
 #include "cache_eviction_method.h"
 #include "caching_method.h"
-#include "fd/pyro/core/profiling_context.h"
+#include "../core/profiling_context.h"
 #include "model/table/column_layout_relation_data.h"
 
 namespace model {
